+++ src/Resources/Resources/ChangeLog.md
@@ -21,7 +21,8 @@
 ## Upcoming Release
 * [Breaking change] Migrate from AAD Graph to Microsoft Graph 
 * Removed breaking change warnings for AAD cmdlets
-  
+* Added new cmdlet `Publish-AzBicepModule` for publishing Bicep modules
+
 ## Version 4.4.1
 * Fixed a bug about the exitcode of Bicep [#16055]
 * Added breaking change warnings for AAD cmdlets